+++ ReleaseNotes.md
@@ -2,6 +2,8 @@ Here's a list of the latest changes for the current release:
 
 **Enhancements**
 
+- [Categories can be marked as online-only now so publication options are not displayed](https://trello.com/c/7P6SKdl6/293-categories-for-online-only-booking)
+
 **Bug Fixes**   
 
 **Change Requests**